:hammer: 현황, 항목 조회 서비스 수정

+++ src/main/java/me/iqpizza6349/dote/domain/vote/service/VoteService.java
@@ -3,8 +3,10 @@ package me.iqpizza6349.dote.domain.vote.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import me.iqpizza6349.dote.domain.member.entity.Member;
+import me.iqpizza6349.dote.domain.member.type.Role;
 import me.iqpizza6349.dote.domain.team.dto.TeamDto;
 import me.iqpizza6349.dote.domain.team.entity.Team;
+import me.iqpizza6349.dote.domain.team.ro.TeamResponseDto;
 import me.iqpizza6349.dote.domain.team.service.TeamService;
 import me.iqpizza6349.dote.domain.vote.dto.BallotDto;
 import me.iqpizza6349.dote.domain.vote.dto.VoteDto;
@@ -15,22 +17,28 @@ import me.iqpizza6349.dote.domain.vote.ro.TeamRO;
 import me.iqpizza6349.dote.domain.vote.ro.VoteRO;
 import org.springframework.data.domain.*;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
 @Slf4j
 @Service
-@Transactional(propagation = Propagation.NESTED)
+@Transactional
 @RequiredArgsConstructor
 public class VoteService {
 
     private final VoteRepository voteRepository;
     private final TeamService teamService;
 
-    public VoteRO createVote(VoteDto voteDto) {
+    public VoteRO createVote(Member member, VoteDto voteDto) {
+        if (isNotAdmin(member.getRole())) {
+            throw new Member.ForbiddenException();
+        }
+
         Set<Team> teamSet = voteDto.getItems()
                 .stream()
                 .map(TeamDto::getName)
@@ -42,8 +50,16 @@ public class VoteService {
         return new VoteRO(voteRepository.save(vote));
     }
     
-    public void deleteVote(long voteId) {
+    public void deleteVote(Member member, long voteId) {
+        if (isNotAdmin(member.getRole())) {
+            throw new Member.ForbiddenException();
+        }
+
         voteRepository.delete(findById(voteId));
+    }
+
+    private boolean isNotAdmin(Role role) {
+        return role != Role.ADMIN;
     }
 
     @Transactional(readOnly = true)
@@ -57,17 +73,28 @@ public class VoteService {
 
     public BallotRO addVote(Member member, long voteId, BallotDto ballotDto) {
         Vote vote = findById(voteId);
+        if (vote.getExpiryDate().isBefore(LocalDateTime.now())) {
+            throw new Vote.AlreadyClosedException();
+        }
+
         teamService.ballot(member, vote, ballotDto.getTeamId());
         return new BallotRO(ballotDto.getTeamId());
     }
 
-    // TODO 현황 조회
     public Page<TeamRO> findStatusInquiry(long voteId, int page) {
         Vote vote = findById(voteId);
         // vote 의 teams 의 현황 조회
         return new PageImpl<>(teamService.findAll(vote, page).stream()
                 .map(TeamRO::new)
                 .collect(Collectors.toList()));
+    }
+
+    public List<TeamResponseDto> findAllTeams(long voteId) {
+        Vote vote = findById(voteId);
+        return teamService.findAllTeams(vote)
+                .stream()
+                .map(TeamResponseDto::new)
+                .collect(Collectors.toCollection(LinkedList::new));
     }
 
     @Transactional(readOnly = true)
